requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
gunicorn
pillow
python-dotenv
django-cors-headers
psycopg2-binary
django-import-export
openpyxl